+++ apps/web/src/lib/config/global-config/constants/constants.ts
