+++ apps/web/src/lib/config/global-config/schema/api-client-schema.ts
@@ -1,15 +1,5 @@
-type apiClientType = {
-    V1: string;
-    V2?: string;
-};
-export type ApiClientsSchemaType = {
-    DASHBOARDS: apiClientType,
-    PROJECT: apiClientType,
-    ASSET_INVENTORY: apiClientType,
-    COST_ANALYSIS: apiClientType,
-    OPS_FLOW: apiClientType,
-    ALERT_MANAGER: apiClientType,
-};
+import type { ApiClientsSchemaType } from '@/lib/config/global-config/types/type';
+
 
 export const ApiClientEndpoint: ApiClientsSchemaType = {
     DASHBOARDS: {
