+++ apps/web/src/services/configurator.ts
@@ -5,7 +5,7 @@ import { isEmpty } from 'lodash';
 import { useMenuStore } from '@/store/menu/menu-store';
 
 import { FeatureSchemaManager } from '@/lib/config/global-config/feature-schema-manager';
-import type { FeatureSchemaType, GlobalServiceConfig } from '@/lib/config/global-config/type';
+import type { FeatureSchemaType, GlobalServiceConfig } from '@/lib/config/global-config/types/type';
 import type { Menu } from '@/lib/menu/config';
 import { DEFAULT_ADMIN_MENU_LIST, DEFAULT_MENU_LIST } from '@/lib/menu/menu-architecture';
